test: migrate zero-secrets to bdd
